@@ -5,7 +5,7 @@
  * 
  * 
  * 
- * import javax.inject.Inject;
+ * import jakarta.inject.Inject;
  * 
  * import org.slf4j.Logger; import org.slf4j.LoggerFactory; import
  * org.springframework.beans.factory.annotation.Autowired; import
