@@ -2,7 +2,7 @@ package com.webapp.site;
 
 import java.util.List;
 
-import javax.validation.constraints.*;
+import jakarta.validation.constraints.*;
 
 public class EventForm {
 
